Add possibility for Wonder Item shuffle

+++ soh/src/overlays/actors/ovl_En_Wonder_Item/z_en_wonder_item.c
@@ -6,6 +6,7 @@
 
 #include "z_en_wonder_item.h"
 #include "vt.h"
+#include "soh/Enhancements/game-interactor/GameInteractor_Hooks.h"
 
 #define FLAGS 0
 
@@ -75,24 +76,26 @@ void EnWonderItem_DropCollectible(EnWonderItem* this, PlayState* play, s32 autoC
     s32 i;
     s32 randomDrop;
 
-    func_80078884(NA_SE_SY_GET_ITEM);
+    if (GameInteractor_Should(VB_WONDER_ITEM_DROP_COLLECTIBLE, true, this)) {
+        func_80078884(NA_SE_SY_GET_ITEM);
 
-    if (this->dropCount == 0) {
-        this->dropCount++;
-    }
-    for (i = this->dropCount; i > 0; i--) {
-        if (this->itemDrop < WONDERITEM_DROP_RANDOM) {
-            if ((this->itemDrop == WONDERITEM_DROP_FLEXIBLE) || !autoCollect) {
-                Item_DropCollectible(play, &this->actor.world.pos, dropTable[this->itemDrop]);
+        if (this->dropCount == 0) {
+            this->dropCount++;
+        }
+        for (i = this->dropCount; i > 0; i--) {
+            if (this->itemDrop < WONDERITEM_DROP_RANDOM) {
+                if ((this->itemDrop == WONDERITEM_DROP_FLEXIBLE) || !autoCollect) {
+                    Item_DropCollectible(play, &this->actor.world.pos, dropTable[this->itemDrop]);
+                } else {
+                    Item_DropCollectible(play, &this->actor.world.pos, dropTable[this->itemDrop] | 0x8000);
+                }
             } else {
-                Item_DropCollectible(play, &this->actor.world.pos, dropTable[this->itemDrop] | 0x8000);
-            }
-        } else {
-            randomDrop = this->itemDrop - WONDERITEM_DROP_RANDOM;
-            if (!autoCollect) {
-                Item_DropCollectibleRandom(play, NULL, &this->actor.world.pos, randomDrop);
-            } else {
-                Item_DropCollectibleRandom(play, NULL, &this->actor.world.pos, randomDrop | 0x8000);
+                randomDrop = this->itemDrop - WONDERITEM_DROP_RANDOM;
+                if (!autoCollect) {
+                    Item_DropCollectibleRandom(play, NULL, &this->actor.world.pos, randomDrop);
+                } else {
+                    Item_DropCollectibleRandom(play, NULL, &this->actor.world.pos, randomDrop | 0x8000);
+                }
             }
         }
     }
